requirements.txt
pandas>=2.2.2
PyMuPDF>=1.24.3
openpyxl>=3.1.2
Pillow>=10.3.0
pytesseract>=0.3.11
colorama>=0.4.6
ollama>=0.1.9
python-dateutil>=2.9.0